+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 * PHP 8.5 Support [@tomasnorre](https://github.com/tomasnorre)
+* TYPO3 14 Support [@tomasnorre](https://github.com/tomasnorre)
 
 ### Changed
 * Updated symfony-components dependencies to ^7.2